refactor: move ConverterCaptureProbe into Core; drop the plugin's friend access (issue #102)
The probe is SimHub-free and used Core internals (Col03FrameClassifier,
WheelEngage) — the sole reason FanaBridge.Core granted
InternalsVisibleTo("FanaBridge"). It now lives in Core/Diagnostics
beside FanatecSoftwareMonitor, and the production friend grant is gone;
the merged build proves the plugin needs no Core internals. Tests keep
their own grant and move to Core/Diagnostics.

748/748 tests green; ILRepack merged build clean.
